BE: [feat] 차터 PDF 재업로드 기능

+++ backend/team-sync/src/main/java/com/pocketstone/team_sync/service/CharterFileService.java
@@ -12,10 +12,7 @@ import com.pocketstone.team_sync.dto.FileUploadResponseDto;
 import com.pocketstone.team_sync.entity.ProjectCharter;
 import com.pocketstone.team_sync.entity.User;
 import com.pocketstone.team_sync.entity.charter.CharterPdf;
-import com.pocketstone.team_sync.exception.CharterDownloadException;
-import com.pocketstone.team_sync.exception.CharterNotFoundException;
-import com.pocketstone.team_sync.exception.CharterUploadException;
-import com.pocketstone.team_sync.exception.ProjectNotFoundException;
+import com.pocketstone.team_sync.exception.*;
 import com.pocketstone.team_sync.repository.ProjectCharterRepository;
 import com.pocketstone.team_sync.repository.charter.CharterPdfRepository;
 import com.pocketstone.team_sync.utility.ProjectValidationUtils;
@@ -30,8 +27,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.StreamUtils;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
-import java.io.FileOutputStream;
+
 import java.io.IOException;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -75,14 +71,19 @@ public class CharterFileService {
         DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
         String todayDate = dateTimeFormatter.format(LocalDate.now());
         String filePath = "";
+
         try {
             ObjectMetadata objectMetadata = new ObjectMetadata();
             objectMetadata.setContentType(multipartFile.getContentType());
             objectMetadata.setContentLength(multipartFile.getSize());
-            filePath = projectCharter.getProject().getProjectName()+"_"+
-                    projectCharter.getProject().getId()+"_"+
-                    projectCharter.getId()+"_"+
+            filePath = projectCharter.getProject().getProjectName() + "_" +
+                    projectCharter.getProject().getId() + "_" +
+                    projectCharter.getId() + "_" +
                     todayDate;
+            boolean isObjectExist = s3Client.doesObjectExist(bucketName, filePath);
+            if (isObjectExist) {
+                throw new CharterPdfAlreadyExistsException();
+            }
             s3Client.putObject(bucketName, filePath, multipartFile.getInputStream(), objectMetadata);
             fileUploadResponse.setFilePath(filePath);
             fileUploadResponse.setDateTime(LocalDateTime.now());
@@ -98,39 +99,59 @@ public class CharterFileService {
         return fileUploadResponse;
     }
 
-    public ResponseEntity<ByteArrayResource> downloadFile(User user, Long projectId ){
+    public ResponseEntity<ByteArrayResource> downloadFile(User user, Long projectId) {
         ProjectCharter projectCharter = projectCharterRepository.findByProjectId(projectId)
                 .orElseThrow(() -> new ProjectNotFoundException(" "));
         ProjectValidationUtils.validateCharterOwner(user, projectCharter);
 
-            CharterPdf charterPdf = charterPdfRepository.findByProjectCharterId(projectCharter.getId())
-                    .orElseThrow(CharterNotFoundException::new);
+        CharterPdf charterPdf = charterPdfRepository.findByProjectCharterId(projectCharter.getId())
+                .orElseThrow(CharterNotFoundException::new);
 
-            try {
-                String filePath = charterPdf.getFileName();
-                boolean isObjectExist = s3Client.doesObjectExist(bucketName, filePath);
-                if (!isObjectExist) {
-                    throw new CharterDownloadException();
-                }
-                S3Object s3Object = s3Client.getObject(bucketName, filePath);
-                S3ObjectInputStream inputStream = s3Object.getObjectContent();
-                byte[] content = StreamUtils.copyToByteArray(inputStream);
-                inputStream.close();
+        try {
+            String filePath = charterPdf.getFileName();
+            boolean isObjectExist = s3Client.doesObjectExist(bucketName, filePath);
+            if (!isObjectExist) {
+                throw new CharterPdfNotFoundException();
+            }
+            S3Object s3Object = s3Client.getObject(bucketName, filePath);
+            S3ObjectInputStream inputStream = s3Object.getObjectContent();
+            byte[] content = StreamUtils.copyToByteArray(inputStream);
+            inputStream.close();
 
-                ByteArrayResource resource = new ByteArrayResource(content);
+            ByteArrayResource resource = new ByteArrayResource(content);
 
-                return ResponseEntity.ok()
-                        .contentLength(content.length)
-                        .contentType(org.springframework.http.MediaType.parseMediaType("application/pdf"))
-                        .header(HttpHeaders.CONTENT_DISPOSITION, "inline; filename=" + filePath)
-                        .body(resource);
-            } catch (IOException e) {
-                throw new RuntimeException(e.getMessage());
+            return ResponseEntity.ok()
+                    .contentLength(content.length)
+                    .contentType(org.springframework.http.MediaType.parseMediaType("application/pdf"))
+                    .header(HttpHeaders.CONTENT_DISPOSITION, "inline; filename=" + filePath)
+                    .body(resource);
+        } catch (IOException e) {
+            throw new RuntimeException(e.getMessage());
+        }
+
+
+    }
+
+    public FileUploadResponseDto reUploadFile(User user, Long projectId, MultipartFile multipartFile) {
+        ProjectCharter projectCharter = projectCharterRepository.findByProjectId(projectId)
+                .orElseThrow(() -> new ProjectNotFoundException(" "));
+        CharterPdf charterPdf = charterPdfRepository.findByProjectCharterId(projectCharter.getId())
+                .orElseThrow(CharterNotFoundException::new);
+        ProjectValidationUtils.validateCharterOwner(user, projectCharter);
+
+
+            boolean isObjectExist = s3Client.doesObjectExist(bucketName, charterPdf.getFileName());
+            if (!isObjectExist) {
+                throw new CharterPdfNotFoundException();
+            } else {
+                s3Client.deleteObject(bucketName, charterPdf.getFileName());
+                charterPdfRepository.delete(charterPdf);
             }
 
+        return uploadFile(user, projectId, multipartFile);
+
 
     }
-
-    }
+}
